+++ zhglxt-web/src/main/java/com/zhglxt/web/controller/system/SysDictTypeController.java
@@ -1,6 +1,7 @@
 package com.zhglxt.web.controller.system;
 
 import com.zhglxt.common.annotation.Log;
+import com.zhglxt.common.config.GlobalConfig;
 import com.zhglxt.common.constant.UserConstants;
 import com.zhglxt.common.core.controller.BaseController;
 import com.zhglxt.common.core.entity.AjaxResult;
@@ -25,13 +26,12 @@ import java.util.List;
 
 /**
  * 数据字典信息
- * 
+ *
  * @author ruoyi
  */
 @Controller
 @RequestMapping("/system/dict")
-public class SysDictTypeController extends BaseController
-{
+public class SysDictTypeController extends BaseController {
     private String prefix = "system/dict/type";
 
     @Autowired
@@ -39,16 +39,14 @@ public class SysDictTypeController extends BaseController
 
     @RequiresPermissions("system:dict:view")
     @GetMapping()
-    public String dictType()
-    {
+    public String dictType() {
         return prefix + "/type";
     }
 
     @PostMapping("/list")
     @RequiresPermissions("system:dict:list")
     @ResponseBody
-    public TableDataInfo list(SysDictType dictType)
-    {
+    public TableDataInfo list(SysDictType dictType) {
         startPage();
         List<SysDictType> list = dictTypeService.selectDictTypeList(dictType);
         return getDataTable(list);
@@ -58,8 +56,7 @@ public class SysDictTypeController extends BaseController
     @RequiresPermissions("system:dict:export")
     @PostMapping("/export")
     @ResponseBody
-    public AjaxResult export(SysDictType dictType)
-    {
+    public AjaxResult export(SysDictType dictType) {
 
         List<SysDictType> list = dictTypeService.selectDictTypeList(dictType);
         ExcelUtil<SysDictType> util = new ExcelUtil<SysDictType>(SysDictType.class);
@@ -70,8 +67,7 @@ public class SysDictTypeController extends BaseController
      * 新增字典类型
      */
     @GetMapping("/add")
-    public String add()
-    {
+    public String add() {
         return prefix + "/add";
     }
 
@@ -82,10 +78,11 @@ public class SysDictTypeController extends BaseController
     @RequiresPermissions("system:dict:add")
     @PostMapping("/add")
     @ResponseBody
-    public AjaxResult addSave(@Validated SysDictType dict)
-    {
-        if (UserConstants.DICT_TYPE_NOT_UNIQUE.equals(dictTypeService.checkDictTypeUnique(dict)))
-        {
+    public AjaxResult addSave(@Validated SysDictType dict) {
+        if (GlobalConfig.isDemoEnabled()) {
+            return error("演示模式不允许本操作");
+        }
+        if (UserConstants.DICT_TYPE_NOT_UNIQUE.equals(dictTypeService.checkDictTypeUnique(dict))) {
             return error("新增字典'" + dict.getDictName() + "'失败，字典类型已存在");
         }
         dict.setCreateBy(getLoginName());
@@ -97,8 +94,7 @@ public class SysDictTypeController extends BaseController
      */
     @RequiresPermissions("system:dict:edit")
     @GetMapping("/edit/{dictId}")
-    public String edit(@PathVariable("dictId") String dictId, ModelMap mmap)
-    {
+    public String edit(@PathVariable("dictId") String dictId, ModelMap mmap) {
         mmap.put("dict", dictTypeService.selectDictTypeById(dictId));
         return prefix + "/edit";
     }
@@ -110,10 +106,11 @@ public class SysDictTypeController extends BaseController
     @RequiresPermissions("system:dict:edit")
     @PostMapping("/edit")
     @ResponseBody
-    public AjaxResult editSave(@Validated SysDictType dict)
-    {
-        if (UserConstants.DICT_TYPE_NOT_UNIQUE.equals(dictTypeService.checkDictTypeUnique(dict)))
-        {
+    public AjaxResult editSave(@Validated SysDictType dict) {
+        if (GlobalConfig.isDemoEnabled()) {
+            return error("演示模式不允许本操作");
+        }
+        if (UserConstants.DICT_TYPE_NOT_UNIQUE.equals(dictTypeService.checkDictTypeUnique(dict))) {
             return error("修改字典'" + dict.getDictName() + "'失败，字典类型已存在");
         }
         dict.setUpdateBy(getLoginName());
@@ -124,8 +121,10 @@ public class SysDictTypeController extends BaseController
     @RequiresPermissions("system:dict:remove")
     @PostMapping("/remove")
     @ResponseBody
-    public AjaxResult remove(String ids)
-    {
+    public AjaxResult remove(String ids) {
+        if (GlobalConfig.isDemoEnabled()) {
+            return error("演示模式不允许本操作");
+        }
         dictTypeService.deleteDictTypeByIds(ids);
         return success();
     }
@@ -137,8 +136,7 @@ public class SysDictTypeController extends BaseController
     @Log(title = "字典类型", businessType = BusinessType.CLEAN)
     @GetMapping("/refreshCache")
     @ResponseBody
-    public AjaxResult refreshCache()
-    {
+    public AjaxResult refreshCache() {
         dictTypeService.resetDictCache();
         return success();
     }
@@ -148,8 +146,7 @@ public class SysDictTypeController extends BaseController
      */
     @RequiresPermissions("system:dict:list")
     @GetMapping("/detail/{dictId}")
-    public String detail(@PathVariable("dictId") String dictId, ModelMap mmap)
-    {
+    public String detail(@PathVariable("dictId") String dictId, ModelMap mmap) {
         mmap.put("dict", dictTypeService.selectDictTypeById(dictId));
         mmap.put("dictList", dictTypeService.selectDictTypeAll());
         return "system/dict/data/data";
@@ -160,8 +157,7 @@ public class SysDictTypeController extends BaseController
      */
     @PostMapping("/checkDictTypeUnique")
     @ResponseBody
-    public String checkDictTypeUnique(SysDictType dictType)
-    {
+    public String checkDictTypeUnique(SysDictType dictType) {
         return dictTypeService.checkDictTypeUnique(dictType);
     }
 
@@ -169,9 +165,7 @@ public class SysDictTypeController extends BaseController
      * 选择字典树
      */
     @GetMapping("/selectDictTree/{columnId}/{dictType}")
-    public String selectDeptTree(@PathVariable("columnId") Long columnId, @PathVariable("dictType") String dictType,
-            ModelMap mmap)
-    {
+    public String selectDeptTree(@PathVariable("columnId") Long columnId, @PathVariable("dictType") String dictType, ModelMap mmap) {
         mmap.put("columnId", columnId);
         mmap.put("dict", dictTypeService.selectDictTypeByType(dictType));
         return prefix + "/tree";
@@ -182,8 +176,7 @@ public class SysDictTypeController extends BaseController
      */
     @GetMapping("/treeData")
     @ResponseBody
-    public List<Ztree> treeData()
-    {
+    public List<Ztree> treeData() {
         List<Ztree> ztrees = dictTypeService.selectDictTree(new SysDictType());
         return ztrees;
     }
