+++ zhglxt-web/src/main/java/com/zhglxt/web/controller/system/SysPostController.java
@@ -1,17 +1,7 @@
 package com.zhglxt.web.controller.system;
 
-import java.util.List;
-import org.apache.shiro.authz.annotation.RequiresPermissions;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.ModelMap;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 import com.zhglxt.common.annotation.Log;
+import com.zhglxt.common.config.GlobalConfig;
 import com.zhglxt.common.constant.UserConstants;
 import com.zhglxt.common.core.controller.BaseController;
 import com.zhglxt.common.core.entity.AjaxResult;
@@ -20,16 +10,23 @@ import com.zhglxt.common.enums.BusinessType;
 import com.zhglxt.common.utils.poi.ExcelUtil;
 import com.zhglxt.system.entity.SysPost;
 import com.zhglxt.system.service.ISysPostService;
+import org.apache.shiro.authz.annotation.RequiresPermissions;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.ModelMap;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 /**
  * 岗位信息操作处理
- * 
+ *
  * @author ruoyi
  */
 @Controller
 @RequestMapping("/system/post")
-public class SysPostController extends BaseController
-{
+public class SysPostController extends BaseController {
     private String prefix = "system/post";
 
     @Autowired
@@ -37,16 +34,14 @@ public class SysPostController extends BaseController
 
     @RequiresPermissions("system:post:view")
     @GetMapping()
-    public String operlog()
-    {
+    public String operlog() {
         return prefix + "/post";
     }
 
     @RequiresPermissions("system:post:list")
     @PostMapping("/list")
     @ResponseBody
-    public TableDataInfo list(SysPost post)
-    {
+    public TableDataInfo list(SysPost post) {
         startPage();
         List<SysPost> list = postService.selectPostList(post);
         return getDataTable(list);
@@ -56,8 +51,7 @@ public class SysPostController extends BaseController
     @RequiresPermissions("system:post:export")
     @PostMapping("/export")
     @ResponseBody
-    public AjaxResult export(SysPost post)
-    {
+    public AjaxResult export(SysPost post) {
         List<SysPost> list = postService.selectPostList(post);
         ExcelUtil<SysPost> util = new ExcelUtil<SysPost>(SysPost.class);
         return util.exportExcel(list, "岗位数据");
@@ -67,14 +61,13 @@ public class SysPostController extends BaseController
     @Log(title = "岗位管理", businessType = BusinessType.DELETE)
     @PostMapping("/remove")
     @ResponseBody
-    public AjaxResult remove(String ids)
-    {
-        try
-        {
+    public AjaxResult remove(String ids) {
+        try {
+            if (GlobalConfig.isDemoEnabled()) {
+                return error("演示模式不允许本操作");
+            }
             return toAjax(postService.deletePostByIds(ids));
-        }
-        catch (Exception e)
-        {
+        } catch (Exception e) {
             return error(e.getMessage());
         }
     }
@@ -83,8 +76,7 @@ public class SysPostController extends BaseController
      * 新增岗位
      */
     @GetMapping("/add")
-    public String add()
-    {
+    public String add() {
         return prefix + "/add";
     }
 
@@ -95,14 +87,13 @@ public class SysPostController extends BaseController
     @Log(title = "岗位管理", businessType = BusinessType.INSERT)
     @PostMapping("/add")
     @ResponseBody
-    public AjaxResult addSave(@Validated SysPost post)
-    {
-        if (UserConstants.POST_NAME_NOT_UNIQUE.equals(postService.checkPostNameUnique(post)))
-        {
-            return error("新增岗位'" + post.getPostName() + "'失败，岗位名称已存在");
+    public AjaxResult addSave(@Validated SysPost post) {
+        if (GlobalConfig.isDemoEnabled()) {
+            return error("演示模式不允许本操作");
         }
-        else if (UserConstants.POST_CODE_NOT_UNIQUE.equals(postService.checkPostCodeUnique(post)))
-        {
+        if (UserConstants.POST_NAME_NOT_UNIQUE.equals(postService.checkPostNameUnique(post))) {
+            return error("新增岗位'" + post.getPostName() + "'失败，岗位名称已存在");
+        } else if (UserConstants.POST_CODE_NOT_UNIQUE.equals(postService.checkPostCodeUnique(post))) {
             return error("新增岗位'" + post.getPostName() + "'失败，岗位编码已存在");
         }
         post.setCreateBy(getLoginName());
@@ -114,8 +105,7 @@ public class SysPostController extends BaseController
      */
     @RequiresPermissions("system:post:edit")
     @GetMapping("/edit/{postId}")
-    public String edit(@PathVariable("postId") Long postId, ModelMap mmap)
-    {
+    public String edit(@PathVariable("postId") String postId, ModelMap mmap) {
         mmap.put("post", postService.selectPostById(postId));
         return prefix + "/edit";
     }
@@ -127,14 +117,13 @@ public class SysPostController extends BaseController
     @Log(title = "岗位管理", businessType = BusinessType.UPDATE)
     @PostMapping("/edit")
     @ResponseBody
-    public AjaxResult editSave(@Validated SysPost post)
-    {
-        if (UserConstants.POST_NAME_NOT_UNIQUE.equals(postService.checkPostNameUnique(post)))
-        {
-            return error("修改岗位'" + post.getPostName() + "'失败，岗位名称已存在");
+    public AjaxResult editSave(@Validated SysPost post) {
+        if (GlobalConfig.isDemoEnabled()) {
+            return error("演示模式不允许本操作");
         }
-        else if (UserConstants.POST_CODE_NOT_UNIQUE.equals(postService.checkPostCodeUnique(post)))
-        {
+        if (UserConstants.POST_NAME_NOT_UNIQUE.equals(postService.checkPostNameUnique(post))) {
+            return error("修改岗位'" + post.getPostName() + "'失败，岗位名称已存在");
+        } else if (UserConstants.POST_CODE_NOT_UNIQUE.equals(postService.checkPostCodeUnique(post))) {
             return error("修改岗位'" + post.getPostName() + "'失败，岗位编码已存在");
         }
         post.setUpdateBy(getLoginName());
@@ -146,8 +135,7 @@ public class SysPostController extends BaseController
      */
     @PostMapping("/checkPostNameUnique")
     @ResponseBody
-    public String checkPostNameUnique(SysPost post)
-    {
+    public String checkPostNameUnique(SysPost post) {
         return postService.checkPostNameUnique(post);
     }
 
@@ -156,8 +144,7 @@ public class SysPostController extends BaseController
      */
     @PostMapping("/checkPostCodeUnique")
     @ResponseBody
-    public String checkPostCodeUnique(SysPost post)
-    {
+    public String checkPostCodeUnique(SysPost post) {
         return postService.checkPostCodeUnique(post);
     }
 }
